fix -v as fVerified was used to prevent verification from running on chunks of a table in subsequent reads - returned in tandem with debug flag to avoid checks for standard runs add cahin walk step and false alarm debug output to standard -v flag move all the stats for chainWalkStep and FalseAlarms to uint64 remove VC++ warning supression pragmas and fix minor warnings that resulted remove unnecessary/redundant debug output

+++ client_applications/rcracki_mt/CrackEngine.h
@@ -104,9 +104,9 @@ public:
 	float GetStatTotalDiskAccessTime();
 	float GetStatTotalCryptanalysisTime();
 	float GetStatTotalPrecalculationTime();
-	uint64   GetStatTotalChainWalkStep();
-	int   GetStatTotalFalseAlarm();
-	uint64   GetStatTotalChainWalkStepDueToFalseAlarm();
+	uint64 GetStatTotalChainWalkStep();
+	uint64 GetStatTotalFalseAlarm();
+	uint64 GetStatTotalChainWalkStepDueToFalseAlarm();
 	void setOutputFile(std::string pathName);
 	void setSession(std::string sSessionPathName, std::string sProgressPathName, std::string sPrecalcPathName, bool keepPrecalc);
 };
